backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
psycopg[binary]==3.1.12
pandas==2.3.2
requests
beautifulsoup4
tqdm
pyyaml
django-cors-headers
djangorestframework